// front-end/src/app/api/models/pokemon-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PokemonDto {

  /**
   * Attack stat
   */
  attack: number;

  /**
   * Defense stat
   */
  defense: number;

  /**
   * Generation number
   */
  generation: number;

  /**
   * HP stat
   */
  hp: number;

  /**
   * Unique identifier for the Pokémon
   */
  id: string;

  /**
   * Image URL of the Pokémon
   */
  image: string;

  /**
   * Is this Pokémon legendary?
   */
  legendary: boolean;

  /**
   * Name of the Pokémon
   */
  name: string;

  /**
   * Special Attack stat
   */
  spAttack: number;

  /**
   * Special Defense stat
   */
  spDefense: number;

  /**
   * Speed stat
   */
  speed: number;

  /**
   * Total base stats
   */
  total: number;

  /**
   * Primary type of the Pokémon
   */
  type1: string;

  /**
   * Secondary type of the Pokémon
   */
  type2?: string;

  /**
   * YouTube URL for the Pokémon
   */
  ytbUrl: string;
}